+++ rapid7/main.py
@@ -37,9 +37,9 @@ import logging
 import logging.handlers
 import threading
 import socket
-import time
+import os
+import traceback
 from jsonpath import jsonpath
-
 from netskope.common.utils import AlertsHelper
 from netskope.integrations.cls.plugin_base import (
     PluginBase,
@@ -64,13 +64,55 @@ from .utils.rapid7_exceptions import (
 from .utils.rapid7_cef_generator import (
     CEFGenerator,
 )
-from .utils.rapid7_ssl import SSLRapid7Handler
+from .utils.rapid7_ssl import SSLSysLogHandler
 
-PLUGIN_NAME = "Rapid7 CLS Plugin"
+PLATFORM_NAME = "Rapid7"
+MODULE_NAME = "CLS"
+PLUGIN_VERSION = "3.0.0"
 
 
 class Rapid7Plugin(PluginBase):
     """The Rapid7 plugin implementation class."""
+
+    def __init__(
+        self,
+        name,
+        *args,
+        **kwargs,
+    ):
+        """Initialize SyslogPlugin class."""
+        super().__init__(
+            name,
+            *args,
+            **kwargs,
+        )
+        self.plugin_name, self.plugin_version = self._get_plugin_info()
+        self.log_prefix = f"{MODULE_NAME} {self.plugin_name} [{name}]"
+
+    def _get_plugin_info(self) -> tuple:
+        """Get plugin name and version from manifest.
+        Returns:
+            tuple: Tuple of plugin's name and version fetched from manifest.
+        """
+        try:
+            file_path = os.path.join(
+                str(os.path.dirname(os.path.abspath(__file__))),
+                "manifest.json",
+            )
+            with open(file_path, "r") as manifest:
+                manifest_json = json.load(manifest)
+                plugin_name = manifest_json.get("name", "Rapid7")
+                plugin_version = manifest_json.get("version", "3.0.0")
+                return (plugin_name, plugin_version)
+        except Exception as exp:
+            self.logger.info(
+                message=(
+                    f"{MODULE_NAME} {PLATFORM_NAME}: Error occurred while"
+                    " getting plugin details. Error: {}".format(exp)
+                ),
+                details=traceback.format_exc(),
+            )
+        return (PLATFORM_NAME, PLUGIN_VERSION)
 
     def get_mapping_value_from_json_path(self, data, json_path):
         """To Fetch the value from given JSON object using given JSON path.
@@ -95,17 +137,19 @@ class Rapid7Plugin(PluginBase):
             fetched value.
         """
         return (
-            data[field]
+            (data[field], True)
             if data[field] or isinstance(data[field], int)
-            else "null"
+            else ("null", False)
         )
 
     def get_subtype_mapping(self, mappings, subtype):
-        """To Retrieve subtype mappings (mappings for subtypes of alerts/events) case insensitively.
+        """To Retrieve subtype mappings (mappings for subtypes
+        of alerts/events) case insensitively.
 
         Args:
             mappings: Mapping JSON from which subtypes are to be retrieved
-            subtype: Subtype (e.g. DLP for alerts) for which the mapping is to be fetched
+            subtype: Subtype (e.g. DLP for alerts) for which
+            the mapping is to be fetched
 
         Returns:
             Fetched mapping JSON object
@@ -117,7 +161,8 @@ class Rapid7Plugin(PluginBase):
             return mappings[subtype.upper()]
 
     def get_headers(self, header_mappings, data, data_type, subtype):
-        """To Create a dictionary of CEF headers from given header mappings for given Netskope alert/event record.
+        """To Create a dictionary of CEF headers from given header
+        mappings for given Netskope alert/event record.
 
         Args:
             subtype: Subtype for which the headers are being transformed
@@ -136,12 +181,16 @@ class Rapid7Plugin(PluginBase):
             mapping_variables = {"$tenant_name": tenant.name}
 
         missing_fields = []
+        mapped_field_flag = False
         # Iterate over mapped headers
         for cef_header, header_mapping in header_mappings.items():
             try:
-                headers[cef_header] = self.get_field_value_from_data(
+                headers[cef_header], mapped_field = self.get_field_value_from_data(
                     header_mapping, data, data_type, subtype, False
                 )
+
+                if mapped_field:
+                    mapped_field_flag = mapped_field
 
                 # Handle variable mappings
                 if (
@@ -154,7 +203,7 @@ class Rapid7Plugin(PluginBase):
             except FieldNotFoundError as err:
                 missing_fields.append(str(err))
 
-        return headers
+        return headers, mapped_field_flag
 
     def get_extensions(self, extension_mappings, data, data_type, subtype):
         """Fetch extensions from given mappings.
@@ -170,33 +219,39 @@ class Rapid7Plugin(PluginBase):
         """
         extension = {}
         missing_fields = []
+        mapped_field_flag = False
 
         # Iterate over mapped extensions
         for cef_extension, extension_mapping in extension_mappings.items():
             try:
-                extension[cef_extension] = self.get_field_value_from_data(
+                extension[cef_extension], mapped_field = self.get_field_value_from_data(
                     extension_mapping,
                     data,
                     data_type,
                     subtype,
                     is_json_path="is_json_path" in extension_mapping,
                 )
+                
+                if mapped_field:
+                    mapped_field_flag = mapped_field
             except FieldNotFoundError as err:
                 missing_fields.append(str(err))
 
-        return extension
+        return extension, mapped_field_flag
 
     def get_field_value_from_data(
         self, extension_mapping, data, data_type, subtype, is_json_path=False
     ):
-        """To Fetch the value of extension based on "mapping" and "default" fields.
+        """To Fetch the value of extension based on
+        "mapping" and "default" fields.
 
         Args:
             extension_mapping: Dict containing "mapping" and "default" fields
             data: Data instance retrieved from Netskope
             subtype: Subtype for which the extension are being transformed
             data_type: Data type for which the headers are being transformed
-            is_json_path: Whether the mapped value is JSON path or direct field name
+            is_json_path: Whether the mapped value is
+            JSON path or direct field name
 
         Returns:
             Fetched values of extension
@@ -216,6 +271,9 @@ class Rapid7Plugin(PluginBase):
            NP    |     NP     |        NP      |           - (Not possible)
         -----------------------------------------------------------------------
         """
+        # mapped_field will be returned as true only if the value returned is\
+        # using the mapping_field and not default_value
+        mapped_field = False
         if (
             "mapping_field" in extension_mapping
             and extension_mapping["mapping_field"]
@@ -227,7 +285,8 @@ class Rapid7Plugin(PluginBase):
                     data, extension_mapping["mapping_field"]
                 )
                 if value:
-                    return ",".join([str(val) for val in value])
+                    mapped_field = True
+                    return ",".join([str(val) for val in value]), mapped_field
                 else:
                     raise FieldNotFoundError(
                         extension_mapping["mapping_field"]
@@ -238,13 +297,24 @@ class Rapid7Plugin(PluginBase):
                 if (
                     extension_mapping["mapping_field"] in data
                 ):  # case #1 and case #4
+                    if (
+                        extension_mapping.get("transformation") == "Time Stamp"
+                        and data[extension_mapping["mapping_field"]]
+                    ):
+                        try:
+                            mapped_field = True
+                            return int(
+                                data[extension_mapping["mapping_field"]]
+                            ), mapped_field
+                        except Exception:
+                            pass
                     return self.get_mapping_value_from_field(
                         data, extension_mapping["mapping_field"]
                     )
                 elif "default_value" in extension_mapping:
                     # If mapped value is not found in response and default
                     # is mapped, map the default value (case #2)
-                    return extension_mapping["default_value"]
+                    return extension_mapping["default_value"], mapped_field
                 else:  # case #6
                     raise FieldNotFoundError(
                         extension_mapping["mapping_field"]
@@ -252,7 +322,7 @@ class Rapid7Plugin(PluginBase):
         else:
             # If mapping is not present, 'default_value' must be there
             # because of validation (case #3 and case #5)
-            return extension_mapping["default_value"]
+            return extension_mapping["default_value"], mapped_field
 
     def map_json_data(self, mappings, data, data_type, subtype):
         """Filter the raw data and returns the filtered data.
@@ -263,7 +333,7 @@ class Rapid7Plugin(PluginBase):
         :return: Mapped data based on fields given in mapping file
         """
 
-        if mappings == []:
+        if mappings == [] or not data:
             return data
 
         mapped_dict = {}
@@ -274,7 +344,8 @@ class Rapid7Plugin(PluginBase):
         return mapped_dict
 
     def transform(self, raw_data: list, data_type: str, subtype: str) -> list:
-        """To Transform the raw netskope JSON data into target platform supported data formats.
+        """To Transform the raw netskope JSON data into
+        target platform supported data formats.
 
         Args:
             raw_data (list): Raw Data.
@@ -293,18 +364,24 @@ class Rapid7Plugin(PluginBase):
                 )
             except KeyError as err:
                 self.logger.error(
-                    f"{PLUGIN_NAME}: Error in rapid7 mapping file. Error: {err}."
+                    "{}: Error in rapid7 mapping file. Error: {}.".format(
+                        self.log_prefix, err
+                    )
                 )
                 raise
             except MappingValidationError as err:
                 self.logger.error(
-                    f"{PLUGIN_NAME}: Mapping Validation Failed. Cause: {err}."
+                    "{}: Mapping Validation Failed. Cause: {}.".format(
+                        self.log_prefix, err
+                    )
                 )
                 raise
             except Exception as err:
                 self.logger.error(
-                    f"{PLUGIN_NAME}: An error occurred while mapping data\
-                        using given json mappings. Error: {err}"
+                    "{}: An error occurred while mapping data\
+                        using given json mappings. Error: {}".format(
+                        self.log_prefix, err
+                    )
                 )
                 raise
 
@@ -314,14 +391,17 @@ class Rapid7Plugin(PluginBase):
                 )
             except KeyError:
                 self.logger.warn(
-                    f"{PLUGIN_NAME}: Subtype {subtype} is not present "
-                    "in the configured mapping file. Transformation of current record "
-                    f"will be skipped."
+                    f"{self.log_prefix}: Subtype {subtype} is not present "
+                    "in the configured mapping file. "
+                    "Transformation of current record will be skipped."
                 )
                 return []
             except Exception as exp:
                 self.logger.error(
-                    message=f"{PLUGIN_NAME}: Error occuured while fetching mapping for data type {data_type} and subtype {subtype}",
+                    message="{}: Error occurred while fetching mapping for "
+                    "data type {} and subtype {}".format(
+                        self.log_prefix, data_type, subtype
+                    ),
                     details=str(exp),
                 )
                 raise
@@ -329,33 +409,34 @@ class Rapid7Plugin(PluginBase):
             transformed_data = []
 
             for data in raw_data:
-                transformed_data.append(
-                    self.map_json_data(
-                        subtype_mapping, data, data_type, subtype
-                    )
+                mapped_dict = self.map_json_data(
+                    subtype_mapping, data, data_type, subtype
                 )
+                if mapped_dict:
+                    transformed_data.append(mapped_dict)
             return transformed_data
 
         else:
-
             try:
                 delimiter, cef_version, rapid7_mappings = get_rapid7_mappings(
                     self.mappings, data_type
                 )
             except KeyError as err:
                 self.logger.error(
-                    f"{PLUGIN_NAME}: Error in rapid7 mapping file. "
+                    f"{self.log_prefix}: Error in rapid7 mapping file. "
                     f"Error: {err}"
                 )
                 raise
             except MappingValidationError as err:
                 self.logger.error(
-                    f"{PLUGIN_NAME}: Mapping Validation Failed. Cause: {err}"
+                    "{}: Mapping Validation Failed. Cause: {}".format(
+                        self.log_prefix, err
+                    )
                 )
                 raise
             except Exception as err:
                 self.logger.error(
-                    f"{PLUGIN_NAME}: An error occurred while mapping data "
+                    f"{self.log_prefix}: An error occurred while mapping data "
                     f"using given json mappings. Error: {err}"
                 )
                 raise
@@ -365,90 +446,98 @@ class Rapid7Plugin(PluginBase):
                 delimiter,
                 cef_version,
                 self.logger,
+                self.log_prefix,
             )
+            
+            # First retrieve the mapping of subtype being transformed
+            try:
+                subtype_mapping = self.get_subtype_mapping(
+                    rapid7_mappings[data_type], subtype
+                )
+            except KeyError:
+                self.logger.warn(
+                    f"{self.log_prefix}: Subtype {subtype} is not present "
+                    "in the configured mapping file. "
+                    "Transformation of current batch will be skipped."
+                )
+                return []
+            except Exception:
+                self.logger.error(
+                    "{}: Error occurred while retrieving mappings "
+                    'for subtype "{}". Transformation of current batch '
+                    "will be skipped.".format(self.log_prefix, subtype)
+                )
+                return []
 
             transformed_data = []
             for data in raw_data:
-
-                # First retrieve the mapping of subtype being transformed
-                try:
-                    subtype_mapping = self.get_subtype_mapping(
-                        rapid7_mappings[data_type], subtype
-                    )
-                except KeyError:
-                    self.logger.warn(
-                        f"{PLUGIN_NAME}: Subtype {subtype} is not present "
-                        "in the configured mapping file. Transformation of current record "
-                        f"will be skipped."
-                    )
-                    return []
-                except Exception:
-                    self.logger.error(
-                        'Error occurred while retrieving mappings for subtype "{}". '
-                        "Transformation of current record will be skipped.".format(
-                            subtype
-                        )
-                    )
+                if not data:
                     continue
 
                 # Generating the CEF header
                 try:
-                    header = self.get_headers(
+                    header, mapped_flag_header = self.get_headers(
                         subtype_mapping["header"], data, data_type, subtype
                     )
                 except Exception as err:
                     self.logger.error(
-                        "{}: [{}][{}]: Error occurred while creating CEF header: {}. Transformation of "
+                        "{}([{}][{}]): Error occurred while creating CEF "
+                        "header: {}. Transformation of "
                         "current record will be skipped.".format(
-                            PLUGIN_NAME, data_type, subtype, str(err)
+                            self.log_prefix, data_type, subtype, err
                         )
                     )
                     continue
 
                 try:
-                    extension = self.get_extensions(
+                    extension, mapped_flag_extension = self.get_extensions(
                         subtype_mapping["extension"], data, data_type, subtype
                     )
                 except Exception as err:
                     self.logger.error(
-                        "{}: [{}][{}]: Error occurred while creating CEF extension: {}. Transformation of "
+                        "{}([{}][{}]): Error occurred while creating CEF "
+                        "extension: {}. Transformation of "
                         "the current record will be skipped".format(
-                            PLUGIN_NAME, data_type, subtype, str(err)
+                            self.log_prefix, data_type, subtype, err
                         )
                     )
                     continue
 
                 try:
-                    transformed_data.append(
-                        cef_generator.get_cef_event(
-                            data,
-                            header,
-                            extension,
-                            data_type,
-                            subtype,
-                            self.configuration.get(
-                                "log_source_identifier", "netskopece"
-                            ),
-                        )
+                    if not (mapped_flag_header or mapped_flag_extension):
+                        continue
+                    cef_generated_event = cef_generator.get_cef_event(
+                        data,
+                        header,
+                        extension,
+                        data_type,
+                        subtype,
+                        self.configuration.get(
+                            "log_source_identifier", "netskopece"
+                        ),
                     )
+                    if cef_generated_event:
+                        transformed_data.append(cef_generated_event)
                 except EmptyExtensionError:
                     self.logger.error(
-                        "{}: [{}][{}]: Got empty extension during transformation."
-                        "Transformation of current record will be skipped".format(
-                            PLUGIN_NAME, data_type, subtype
+                        "{}([{}][{}]): Got empty extension during "
+                        "transformation. Transformation of current record "
+                        "will be skipped".format(
+                            self.log_prefix, data_type, subtype
                         )
                     )
                 except Exception as err:
                     self.logger.error(
-                        "{}: [{}][{}]: An error occurred during transformation."
-                        " Error: {}".format(
-                            PLUGIN_NAME, data_type, subtype, str(err)
+                        "{}([{}][{}]): An error occurred during "
+                        "transformation. Error: {}".format(
+                            self.log_prefix, data_type, subtype, err
                         )
                     )
             return transformed_data
 
     def init_handler(self, configuration):
-        """Initialize unique Rapid7 handler per thread based on configured protocol."""
+        """Initialize unique Rapid7 handler per thread
+        based on configured protocol."""
         syslogger = logging.getLogger(
             "SYSLOG_LOGGER_{}".format(threading.get_ident())
         )
@@ -457,7 +546,9 @@ class Rapid7Plugin(PluginBase):
         syslogger.propagate = False
 
         if configuration["rapid7_protocol"] == "TLS":
-            tls_handler = SSLRapid7Handler(
+            tls_handler = SSLSysLogHandler(
+                configuration.get("transformData", True),
+                configuration["rapid7_protocol"],
                 address=(
                     configuration["rapid7_server"],
                     configuration["rapid7_port"],
@@ -471,7 +562,9 @@ class Rapid7Plugin(PluginBase):
                 socktype = socket.SOCK_STREAM
 
             # Create a rapid7 handler with given configuration parameters
-            handler = logging.handlers.SysLogHandler(
+            handler = SSLSysLogHandler(
+                configuration.get("transformData", True),
+                configuration["rapid7_protocol"],
                 address=(
                     configuration["rapid7_server"],
                     configuration["rapid7_port"],
@@ -480,10 +573,13 @@ class Rapid7Plugin(PluginBase):
             )
 
             if configuration["rapid7_protocol"] == "TCP":
-                # This will add a line break to the message before it is 'emitted' which ensures that the messages are
-                # split up over multiple lines, see https://bugs.python.org/issue28404
+                # This will add a line break to the message before
+                # it is 'emitted' which ensures that the messages are
+                # split up over multiple lines,
+                # see https://bugs.python.org/issue28404
                 handler.setFormatter(logging.Formatter("%(message)s\n"))
-                # In order for the above to work, then we need to ensure that the null terminator is not included
+                # In order for the above to work, then we need to ensure that
+                # the null terminator is not included
                 handler.append_nul = False
 
             syslogger.addHandler(handler)
@@ -496,7 +592,7 @@ class Rapid7Plugin(PluginBase):
             syslogger = self.init_handler(self.configuration)
         except Exception as err:
             self.logger.error(
-                f"{PLUGIN_NAME}: Error occurred during initializing "
+                f"{self.log_prefix}: Error occurred during initializing "
                 f"connection. Error: {err}"
             )
             raise
@@ -504,12 +600,17 @@ class Rapid7Plugin(PluginBase):
         # Log the transformed data to given rapid7 server
         for data in transformed_data:
             try:
-                syslogger.info(data)
-                syslogger.handlers[0].flush()
+                if data:
+                    syslogger.info(
+                        json.dumps(data) if isinstance(data, dict) else data
+                    )
+                    syslogger.handlers[0].flush()
             except Exception as err:
                 self.logger.error(
-                    f"{PLUGIN_NAME}: Error occurred during data ingestion. "
-                    f"Error: {err}. Record will be skipped"
+                    "{}: Error occurred during data ingestion. "
+                    "Error: {}. Record will be skipped".format(
+                        self.log_prefix, err
+                    )
                 )
 
         # Clean up
@@ -519,7 +620,9 @@ class Rapid7Plugin(PluginBase):
             del syslogger
         except Exception as err:
             self.logger.error(
-                f"{PLUGIN_NAME}: Error occurred during Clean up. Error: {err}"
+                "{}: Error occurred during Clean up. Error: {}".format(
+                    self.log_prefix, err
+                )
             )
 
     def test_server_connectivity(self, configuration):
@@ -528,9 +631,9 @@ class Rapid7Plugin(PluginBase):
             syslogger = self.init_handler(configuration)
         except Exception as err:
             self.logger.error(
-                "Error occurred while establishing connection with rapid7 "
-                "server. Make sure you have provided correct rapid7 "
-                "server and port."
+                f"{self.log_prefix}: Error occurred while establishing "
+                "connection with rapid7 server. Make sure you have provided "
+                "correct rapid7 server and port."
             )
             raise err
         else:
@@ -542,21 +645,21 @@ class Rapid7Plugin(PluginBase):
 
     def validate(self, configuration: dict) -> ValidationResult:
         """Validate the configuration parameters dict."""
-        rapid7_validator = Rapid7Validator(self.logger)
+        rapid7_validator = Rapid7Validator(self.logger, self.log_prefix)
         if (
             "rapid7_server" not in configuration
             or not configuration["rapid7_server"].strip()
         ):
             err_msg = "Rapid7 Server IP/FQDN is a required field"
             self.logger.error(
-                f"{PLUGIN_NAME}: Validation error occurred. Error: "
+                f"{self.log_prefix}: Validation error occurred. Error: "
                 f"{err_msg} in the "
                 "configuration parameters."
             )
             return ValidationResult(success=False, message=err_msg)
         elif type(configuration["rapid7_server"]) != str:
             self.logger.error(
-                f"{PLUGIN_NAME}: Validation error occurred. Error: "
+                f"{self.log_prefix}: Validation error occurred. Error: "
                 "Invalid Rapid7 Server IP/FQDN found in the "
                 "configuration parameters."
             )
@@ -569,7 +672,7 @@ class Rapid7Plugin(PluginBase):
         ):
             err_msg = "Rapid7 Format is a required field"
             self.logger.error(
-                f"{PLUGIN_NAME} Validation error occurred. Error: "
+                f"{self.log_prefix}: Validation error occurred. Error: "
                 f"{err_msg} in the "
                 "configuration parameters."
             )
@@ -579,7 +682,7 @@ class Rapid7Plugin(PluginBase):
             or configuration["rapid7_format"] not in SYSLOG_FORMATS
         ):
             self.logger.error(
-                f"{PLUGIN_NAME}: Validation error occurred. Error: "
+                f"{self.log_prefix}: Validation error occurred. Error: "
                 "Invalid Rapid7 Format found in the configuration parameters."
             )
             return ValidationResult(
@@ -591,7 +694,7 @@ class Rapid7Plugin(PluginBase):
         ):
             err_msg = "Rapid7 Protocol is a required field"
             self.logger.error(
-                f"{PLUGIN_NAME}: Validation error occurred. Error: "
+                f"{self.log_prefix}: Validation error occurred. Error: "
                 f"{err_msg} in the "
                 "configuration parameters."
             )
@@ -601,7 +704,7 @@ class Rapid7Plugin(PluginBase):
             or configuration["rapid7_protocol"] not in SYSLOG_PROTOCOLS
         ):
             self.logger.error(
-                f"{PLUGIN_NAME}: Validation error occurred. Error: "
+                f"{self.log_prefix}: Validation error occurred. Error: "
                 "Invalid Rapid7 Protocol provided in the "
                 "configuration parameters."
             )
@@ -614,7 +717,7 @@ class Rapid7Plugin(PluginBase):
         ):
             err_msg = "Rapid7 Port is a required field"
             self.logger.error(
-                f"{PLUGIN_NAME}: Validation error occurred. Error: "
+                f"{self.log_prefix}: Validation error occurred. Error: "
                 f"{err_msg} in the "
                 "configuration parameters."
             )
@@ -623,7 +726,7 @@ class Rapid7Plugin(PluginBase):
             configuration["rapid7_port"]
         ):
             self.logger.error(
-                f"{PLUGIN_NAME}: Validation error occurred. Error: "
+                f"{self.log_prefix}: Validation error occurred. Error: "
                 "Invalid Rapid7 Port found in the configuration parameters."
             )
             return ValidationResult(
@@ -637,7 +740,7 @@ class Rapid7Plugin(PluginBase):
         ):
             err_msg = "Invalid Rapid7 attribute mapping"
             self.logger.error(
-                f"{PLUGIN_NAME}: Validation error occurred. Error: "
+                f"{self.log_prefix}: Validation error occurred. Error: "
                 f"{err_msg} found in the "
                 "configuration parameters."
             )
@@ -649,9 +752,10 @@ class Rapid7Plugin(PluginBase):
             "rapid7_certificate" not in configuration
             or not configuration["rapid7_certificate"].strip()
         ):
-            err_msg = "Rapid7 Certificate mapping is a required field when TLS is provided"
+            err_msg = "Rapid7 Certificate mapping is a "
+            "required field when TLS is provided"
             self.logger.error(
-                f"{PLUGIN_NAME}: Validation error occurred. Error: "
+                f"{self.log_prefix}: Validation error occurred. Error: "
                 f"{err_msg} in the configuration parameters."
             )
             return ValidationResult(
@@ -664,7 +768,7 @@ class Rapid7Plugin(PluginBase):
         ):
             err_msg = "Invalid Rapid7 Certificate mapping"
             self.logger.error(
-                f"{PLUGIN_NAME}: Validation error occurred. Error: "
+                f"{self.log_prefix}: Validation error occurred. Error: "
                 f"{err_msg} found in the "
                 "configuration parameters."
             )
@@ -678,7 +782,7 @@ class Rapid7Plugin(PluginBase):
         ):
             err_msg = "Log Source Identifier is a required field"
             self.logger.error(
-                f"{PLUGIN_NAME}: Validation error occurred. Error: "
+                f"{self.log_prefix}: Validation error occurred. Error: "
                 f"{err_msg} in the "
                 "configuration parameters."
             )
@@ -690,9 +794,12 @@ class Rapid7Plugin(PluginBase):
             type(configuration["log_source_identifier"]) != str
             or " " in configuration["log_source_identifier"].strip()
         ):
-            err_msg = "Invalid Log Source Identifier provided in the configuration parameters."
+            err_msg = "Invalid Log Source Identifier provided in "
+            "the configuration parameters."
             self.logger.error(
-                f"{PLUGIN_NAME}: Validation error occurred. Error: {err_msg}"
+                "{}: Validation error occurred. Error: {}".format(
+                    self.log_prefix, err_msg
+                )
             )
             return ValidationResult(
                 success=False,
@@ -703,7 +810,7 @@ class Rapid7Plugin(PluginBase):
             self.test_server_connectivity(configuration)
         except Exception:
             self.logger.error(
-                f"{PLUGIN_NAME}: Validation error occurred. Error: "
+                f"{self.log_prefix}: Validation error occurred. Error: "
                 "Connection to SIEM platform is not established."
             )
             return ValidationResult(