+++ rapid7/utils/rapid7_validator.py
@@ -42,9 +42,10 @@ from jsonschema.exceptions import ValidationError as JsonSchemaValidationError
 class Rapid7Validator(object):
     """Rapid7 validator class."""
 
-    def __init__(self, logger):
+    def __init__(self, logger, log_prefix):
         """Initialize."""
         super().__init__()
+        self.log_prefix = log_prefix
         self.logger = logger
 
     def validate_rapid7_port(self, rapid7_port):
@@ -54,7 +55,8 @@ class Rapid7Validator(object):
             rapid7_port: the rapid7 port to be validated
 
         Returns:
-            Whether the provided value is valid or not. True in case of valid value, False otherwise
+            Whether the provided value is valid or not.
+            True in case of valid value, False otherwise
         """
         if rapid7_port or rapid7_port == 0:
             try:
@@ -105,7 +107,7 @@ class Rapid7Validator(object):
                         ".*": {
                             "type": "array",
                         }
-                    }
+                    },
                 },
             },
         }
@@ -143,8 +145,10 @@ class Rapid7Validator(object):
             validate(instance=mappings, schema=schema)
         except JsonSchemaValidationError as err:
             self.logger.error(
-                "Rapid7 Plugin: Validation error occurred. Error: "
-                "validating JSON schema: {}".format(err)
+                "{}: Validation error occurred. "
+                "Error: validating JSON schema: {}".format(
+                    self.log_prefix, err
+                )
             )
             return False
 
@@ -158,9 +162,11 @@ class Rapid7Validator(object):
                         self.validate_taxonomy(subtype_taxonomy)
                     except JsonSchemaValidationError as err:
                         self.logger.error(
-                            "Rapid7 Plugin: Validation error occurred. Error: "
-                            'while validating JSON schema for type "{}" and subtype "{}": '
-                            "{}".format(data_type, subtype, err)
+                            "{}: Validation error occurred. "
+                            "Error: while validating JSON schema for "
+                            'type "{}" and subtype "{}": "{}"'.format(
+                                self.log_prefix, data_type, subtype, err
+                            )
                         )
                         return False
         return True
@@ -172,7 +178,8 @@ class Rapid7Validator(object):
             mappings: the JSON string to be validated
 
         Returns:
-            Whether the provided value is valid or not. True in case of valid value, False otherwise
+            Whether the provided value is valid or not.
+            True in case of valid value, False otherwise
         """
         if mappings is None:
             return False
@@ -181,8 +188,8 @@ class Rapid7Validator(object):
                 return True
         except Exception as err:
             self.logger.error(
-                "Rapid7 Plugin: Validation error occurred. Error: {}".format(
-                    str(err)
+                "{}: Validation error occurred. Error: {}".format(
+                    self.log_prefix, err
                 )
             )
 
@@ -195,7 +202,8 @@ class Rapid7Validator(object):
             valid_extensions: the CSV string to be validated
 
         Returns:
-            Whether the provided value is valid or not. True in case of valid value, False otherwise
+            Whether the provided value is valid or not.
+            True in case of valid value, False otherwise
         """
         try:
             csviter = csv.DictReader(